Add lint and check gates for airship-utils
This is base CR for all other CRs in airship-utils.
It sets up gates config with .zuul.yaml.

It was intended that airship-utils project is a collection of utils
can be used in Airship. At the moment only one util is available -
miniMirror. So all gates checking aims to miniMirror checks.

Two gates available:

- Lint gate runs shellcheck on assets/*.sh
  and hadolint Dockerfile. Can be checked locally with "make lint".
- Check gate builds an image and runs container from the image.
  It also checks that it's possible to install Debian packages from
  the running container.

Change-Id: I354cf96a4ddcd9f1f077a4a93869178ec6c366ab